requirements.txt
dropbox==12.0.2
face_recognition==1.3.0
Flask==3.1.0
opencv_python==4.10.0.84
python-dotenv==1.1.0
Requests==2.32.3
